+++ web/app/themes/brocooly/src/Http/Controllers/PageController.php
@@ -11,8 +11,6 @@ declare(strict_types=1);
 
 namespace Theme\Http\Controllers;
 
-use Theme\Brocooly;
-
 class PageController
 {